+++ src/main/kotlin/ru/yoomoney/gradle/plugins/release/ReleasePlugin.kt
@@ -72,7 +72,7 @@ class ReleasePlugin : Plugin<Project> {
             preReleaseRotateVersionTask.gitSettings = gitSettings
             preReleaseRotateVersionTask.pullRequestLinkSettings = PullRequestLinkSettings(
                     pullRequestLinkInChangelogEnabled = releaseExtension.addPullRequestLinkToChangelog,
-                    pullRequestInfoProvider = PullRequestInfoProvider.valueOf(releaseExtension.pullRequestInfoProvider),
+                    pullRequestInfoProvider = PullRequestInfoProvider.fromName(releaseExtension.pullRequestInfoProvider)!!,
                     bitbucketUser = releaseExtension.bitbucketUser,
                     bitbucketPassword = releaseExtension.bitbucketPassword,
                     githubAccessToken = releaseExtension.githubAccessToken